+++ src/components/Row.js
@@ -88,6 +88,7 @@ const rowFunctionComponent = (props) => {
       shouldShowSelectIcon,
       styleTheme,
       is_selected,
+      renderStyle
    } = props;
    const [hasScrolled, setHasScrolled] = React.useState(false);
    const [rowRef, setRowRef] = React.useState(null);
@@ -382,13 +383,14 @@ const rowFunctionComponent = (props) => {
       // cDP(
       <tr
          ref={scrollTo}
-         className={className + ' tr_shaded'}
+         className={className + ' tr_shaded custom-virtualized-row'}
          onClick={props.onRowClick ? onRowClick(row) : undefined}
          key={rowIndex}
-         style={getItemStyle(
-            isDragging,
-            null
-         )}
+         style={{...getItemStyle(
+               isDragging,
+               null
+            ),
+         }}
       >
          {isDragColumnVisible && cDS(
             <td